# Cache frames for fast reuse.