Rather than allocating a frame on every call, we keep an already allocated list of frames ready to go. This trades lightly more memory usage for faster calls.

This can have a dramatic improvement. On `bm_call_simple` the execution time goes from ~851ms to ~281ms for almost a 67% reduction (CPython is ~10ms - yes ten). Other benchmarks are similarly improved when they make a number of calls:

```
benchmark                          old ns/op     new ns/op     delta
BenchmarkCallSimple                851509025     281303472     -66.96%
BenchmarkCallDefaults              259           118           -54.44%
BenchmarkCallKeywords              461           302           -34.49%
BenchmarkCallKwargs                1449          1252          -13.60%
BenchmarkCallNoArgs                254           105           -58.66%
BenchmarkCallPositionalArgs        284           121           -57.39%
BenchmarkCallVarArgs               497           341           -31.39%
BenchmarkDictCompCreate            479559        466262        -2.77%
BenchmarkGeneratorExpCreate        1335          1324          -0.82%
BenchmarkGeneratorExpIterate       128           132           +3.12%
BenchmarkListCompCreate            133363        132768        -0.45%
BenchmarkArithmetic                1010          916           -9.31%
BenchmarkArithmeticParallel10      1057          1001          -5.30%
BenchmarkArithmeticParallel11      1034          989           -4.35%
BenchmarkArithmeticParallel12      1083          1002          -7.48%
BenchmarkArithmeticParallel2       1047          1006          -3.92%
BenchmarkArithmeticParallel3       1058          1019          -3.69%
BenchmarkArithmeticParallel5       1051          1006          -4.28%
BenchmarkArithmeticParallel6       1063          976           -8.18%
BenchmarkArithmeticParallel7       1065          993           -6.76%
BenchmarkArithmeticParallel8       1052          995           -5.42%
BenchmarkArithmeticParallel9       1070          1003          -6.26%
BenchmarkFib                       15733         5572          -64.58%
BenchmarkFibParallel10             18298         5512          -69.88%
BenchmarkFibParallel11             19196         5458          -71.57%
BenchmarkFibParallel12             19903         5604          -71.84%
BenchmarkFibParallel2              15884         5433          -65.80%
BenchmarkFibParallel3              16013         5402          -66.26%
BenchmarkFibParallel4              16348         5447          -66.68%
BenchmarkFibParallel5              17124         5454          -68.15%
BenchmarkFibParallel6              16710         5425          -67.53%
BenchmarkFibParallel7              17466         5432          -68.90%
BenchmarkFibParallel8              17853         5534          -69.00%
BenchmarkFibParallel9              18187         5460          -69.98%
BenchmarkDictGetItem               137           131           -4.38%
BenchmarkDictStringOnlyGetItem     138           134           -2.90%
BenchmarkDictStringOnlySetItem     274           266           -2.92%
BenchmarkHashStrCached             118           114           -3.39%
BenchmarkGeneratorCreate           466           508           +9.01%
BenchmarkListContains10            1000          1007          +0.70%
BenchmarkListContains3             488           484           -0.82%
BenchmarkListGetItem               109           110           +0.92%
BenchmarkForXRange                 74.4          78.2          +5.11%
BenchmarkWhileCounter              486           489           +0.62%
BenchmarkWhileXRange               428           445           +3.97%
BenchmarkTupleContains10           2448          2415          -1.35%
BenchmarkTupleContains3            1030          1058          +2.72%
BenchmarkTupleGetItem              101           100           -0.99%

benchmark                         old allocs     new allocs     delta
BenchmarkCallSimple               3368434        6              -100.00%
BenchmarkCallDefaults             1              0              -100.00%
BenchmarkCallKeywords             2              1              -50.00%
BenchmarkCallKwargs               8              7              -12.50%
BenchmarkCallNoArgs               1              0              -100.00%
BenchmarkCallPositionalArgs       1              0              -100.00%
BenchmarkCallVarArgs              3              2              -33.33%
BenchmarkArithmetic               10             9              -10.00%
BenchmarkArithmeticParallel10     10             9              -10.00%
BenchmarkArithmeticParallel11     10             9              -10.00%
BenchmarkArithmeticParallel12     10             9              -10.00%
BenchmarkArithmeticParallel2      10             9              -10.00%
BenchmarkArithmeticParallel3      10             9              -10.00%
BenchmarkArithmeticParallel4      10             9              -10.00%
BenchmarkArithmeticParallel5      10             9              -10.00%
BenchmarkArithmeticParallel6      10             9              -10.00%
BenchmarkArithmeticParallel7      10             9              -10.00%
BenchmarkArithmeticParallel8      10             9              -10.00%
BenchmarkArithmeticParallel9      10             9              -10.00%
BenchmarkFib                      187            10             -94.65%
BenchmarkFibParallel10            187            10             -94.65%
BenchmarkFibParallel11            187            10             -94.65%
BenchmarkFibParallel12            187            10             -94.65%
BenchmarkFibParallel2             187            10             -94.65%
BenchmarkFibParallel3             187            10             -94.65%
BenchmarkFibParallel4             187            10             -94.65%
BenchmarkFibParallel5             187            10             -94.65%
BenchmarkFibParallel6             187            10             -94.65%
BenchmarkFibParallel7             187            10             -94.65%
BenchmarkFibParallel8             187            10             -94.65%
BenchmarkFibParallel9             187            10             -94.65%

benchmark                         old bytes     new bytes     delta
BenchmarkCallSimple               323369256     432           -100.00%
BenchmarkCallDefaults             127           31            -75.59%
BenchmarkCallKeywords             207           111           -46.38%
BenchmarkCallKwargs               479           383           -20.04%
BenchmarkCallNoArgs               127           31            -75.59%
BenchmarkCallPositionalArgs       127           31            -75.59%
BenchmarkCallVarArgs              207           111           -46.38%
BenchmarkDictCompCreate           142045        142061        +0.01%
BenchmarkGeneratorExpCreate       655           671           +2.44%
BenchmarkListCompCreate           40367         40383         +0.04%
BenchmarkArithmetic               352           256           -27.27%
BenchmarkArithmeticParallel10     352           256           -27.27%
BenchmarkArithmeticParallel11     352           256           -27.27%
BenchmarkArithmeticParallel12     352           256           -27.27%
BenchmarkArithmeticParallel2      352           256           -27.27%
BenchmarkArithmeticParallel3      352           256           -27.27%
BenchmarkArithmeticParallel4      352           256           -27.27%
BenchmarkArithmeticParallel5      352           256           -27.27%
BenchmarkArithmeticParallel6      352           256           -27.27%
BenchmarkArithmeticParallel7      352           256           -27.27%
BenchmarkArithmeticParallel8      352           256           -27.27%
BenchmarkArithmeticParallel9      352           256           -27.27%
BenchmarkFib                      17328         336           -98.06%
BenchmarkFibParallel10            17330         337           -98.06%
BenchmarkFibParallel11            17331         337           -98.06%
BenchmarkFibParallel12            17331         337           -98.06%
BenchmarkFibParallel2             17328         336           -98.06%
BenchmarkFibParallel3             17328         336           -98.06%
BenchmarkFibParallel4             17329         336           -98.06%
BenchmarkFibParallel5             17329         336           -98.06%
BenchmarkFibParallel6             17329         336           -98.06%
BenchmarkFibParallel7             17329         336           -98.06%
BenchmarkFibParallel8             17330         336           -98.06%
BenchmarkFibParallel9             17330         337           -98.06%
BenchmarkGeneratorCreate          231           247           +6.93%
```